[wrangler] Move vectorize/ commands to src/commands/vectorize/ (Phase 2f)
