+++ packages/wrangler/src/commands/vectorize/createMetadataIndex.ts
@@ -1,5 +1,5 @@
-import { createCommand } from "../core/create-command";
-import { logger } from "../logger";
+import { createCommand } from "../../core/create-command";
+import { logger } from "../../logger";
 import { createMetadataIndex } from "./client";
 import type {
 	VectorizeMetadataIndexProperty,
