+++ packages/wrangler/src/commands/vectorize/delete.ts
@@ -1,6 +1,6 @@
-import { createCommand } from "../core/create-command";
-import { confirm } from "../dialogs";
-import { logger } from "../logger";
+import { createCommand } from "../../core/create-command";
+import { confirm } from "../../dialogs";
+import { logger } from "../../logger";
 import { deleteIndex } from "./client";
 import { deprecatedV1DefaultFlag } from "./common";
