+++ packages/wrangler/src/commands/vectorize/getByIds.ts
@@ -1,6 +1,6 @@
 import { UserError } from "@cloudflare/workers-utils";
-import { createCommand } from "../core/create-command";
-import { logger } from "../logger";
+import { createCommand } from "../../core/create-command";
+import { logger } from "../../logger";
 import { getByIds } from "./client";
 import type { VectorizeVectorIds } from "./types";
