+++ packages/wrangler/src/commands/vectorize/info.ts
@@ -1,5 +1,5 @@
-import { createCommand } from "../core/create-command";
-import { logger } from "../logger";
+import { createCommand } from "../../core/create-command";
+import { logger } from "../../logger";
 import { indexInfo } from "./client";
 
 export const vectorizeInfoCommand = createCommand({
