+++ packages/wrangler/src/commands/vectorize/types.ts
